ASoC: msm: qdsp6v2: Fix improper channel config for aac
AAC encode is failing in ADSP due to mismatch
in channel config datatype.
Update channel config data type for aac encoder.

+++ include/sound/apr_audio-v2.h
@@ -2943,7 +2943,7 @@ struct asm_aac_enc_cfg_v2_t {
 	 * number of channels at the input.
 	 * The number of channels must not change during encoding.
 	 */
-	uint32_t     channel_cfg;
+	uint16_t     channel_cfg;
 
 	/*
 	 * Number of samples per second.
